# apps/gateway/src/swagger/swagger.doc.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/signy-email/create-email-template": {
      "post": {
        "operationId": "SignyEmailController_createEmailTemplate",
        "summary": "Create template for email",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmailTemplateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTemplateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Signy-email"
        ],
        "security": [
          {
            "X_SESSION_KEY": []
          },
          {
            "X_API_KEY": []
          }
        ]
      }
    },
    "/api/signy-email/upload-template-files": {
      "post": {
        "operationId": "SignyEmailController_uploadTemplateFilesToS3",
        "summary": "Upload email template files to S3",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UploadTemplateFilesToS3Request"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadTemplateFilesToS3Response"
                }
              }
            }
          },
          "400": {
            "description": "Bad response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Signy-email"
        ],
        "security": [
          {
            "X_SESSION_KEY": []
          },
          {
            "X_API_KEY": []
          }
        ]
      }
    },
    "/api/signy-email/set-company-logo": {
      "post": {
        "operationId": "SignyEmailController_uploadCompanyLogo",
        "summary": "Set company logo in template",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UploadSignyCompanyLogoRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadSignyCompanyLogoResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Signy-email"
        ],
        "security": [
          {
            "X_SESSION_KEY": []
          },
          {
            "X_API_KEY": []
          }
        ]
      }
    },
    "/api/signy-email/get-email-template-by-id": {
      "post": {
        "operationId": "SignyEmailController_getSignyEmailTemplateById",
        "summary": "Get template by id with permission",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetEmailTemplateWithPermissionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSignyEmailTemplateByIdResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Signy-email"
        ],
        "security": [
          {
            "X_SESSION_KEY": []
          },
          {
            "X_API_KEY": []
          }
        ]
      }
    },
    "/api/signy-email/send-email": {
      "post": {
        "operationId": "SignyEmailController_sendEmail",
        "summary": "Send email",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignySendEmailRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Signy-email"
        ],
        "security": [
          {
            "X_SESSION_KEY": []
          },
          {
            "X_API_KEY": []
          }
        ]
      }
    },
    "/api/signy-email/send-sms": {
      "post": {
        "operationId": "SignyEmailController_sendSms",
        "summary": "Send sms",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignySendSmsRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendSmsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Signy-email"
        ],
        "security": [
          {
            "X_SESSION_KEY": []
          },
          {
            "X_API_KEY": []
          }
        ]
      }
    },
    "/api/signy-email/get-sms-template-by-id": {
      "post": {
        "operationId": "SignyEmailController_getSignySmsTemplateById",
        "summary": "Get template by id with permission",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetSmsTemplateWithPermissionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSignySmsTemplateByIdResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Signy-email"
        ],
        "security": [
          {
            "X_SESSION_KEY": []
          },
          {
            "X_API_KEY": []
          }
        ]
      }
    },
    "/api/signy-email/create-sms-template": {
      "post": {
        "operationId": "SignyEmailController_createSmsTemplate",
        "summary": "Create template for sms",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSmsTemplateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSmsTemplateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Signy-email"
        ],
        "security": [
          {
            "X_SESSION_KEY": []
          },
          {
            "X_API_KEY": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "",
    "description": "\"Rentil\" - API documentation.<br><br>Link to <a  target='_blank' href=''>API Guide</a>",
    "version": "1.1.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "{schema}://undefined",
      "description": "API Server",
      "variables": {
        "schema": {
          "enum": [
            "https"
          ],
          "default": "https"
        }
      }
    },
    {
      "url": "{schema}://localhost:3000",
      "description": "Local API Server",
      "variables": {
        "schema": {
          "enum": [
            "http"
          ],
          "default": "http"
        }
      }
    }
  ],
  "components": {
    "securitySchemes": {
      "X_API_KEY": {
        "type": "apiKey",
        "name": "X-API-Key",
        "description": "Application key",
        "in": "header"
      },
      "X_SESSION_KEY": {
        "type": "apiKey",
        "name": "X-Session-Key",
        "description": "Application key",
        "in": "header"
      }
    },
    "schemas": {
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Message for error alert"
          },
          "title": {
            "type": "string",
            "description": "Title for error alert"
          },
          "errorCode": {
            "type": "integer",
            "description": "1: AccessDenied<br>2: WrongAppKey<br>3: InternalServerError<br>4: WrongInput<br>5: UserNotFoundByEmail<br>6: UserNotFound<br>7: IncorrectEmailOrPassword<br>8: EmailAlreadyRegistered<br>9: UserCredentialNotFound<br>10: EmailAlreadyVerified<br>11: PasswordNotMatch<br>12: OTPCodeExpired<br>13: WrongPasswordRecoveryToken<br>14: PasswordRecoveryTokenExpired<br>15: UserEmailIsNotVerified<br>16: UserPhoneIsNotVerified<br>17: AccountInactive<br>18: ImageFileRequired<br>19: WrongImageFormat<br>20: FileTooLarge<br>21: EmailWrong<br>22: NotOwnerOfAssignmentOrAccepted<br>23: AssignmentNotFound<br>24: NewsDeletedOrAccepted<br>25: SubmissionNotFound<br>26: ExceededLimitOfFiles<br>27: NewPasswordEqualCurrent<br>28: NewsChannelNotFound<br>29: NewsNotFound<br>30: NewsRoomNotFound<br>31: AlreadyPublished<br>32: InvitationNotFound<br>33: PasswordExists<br>34: StaffNotFoundByEmail<br>35: StaffEmailIsNotVerified<br>36: OTPInvitaionCodeExpired<br>37: StaffCredentialNotFound<br>38: StaffNotFound<br>39: InvitationAlreadyExists<br>40: MediaTypeNotAllowed<br>41: DeletedUserAccount<br>42: WrongPassword<br>43: NotAdmin<br>44: NoFilesInSubmission<br>45: VideoTooLong<br>46: SubmissionNotAccepted<br>47: UserNameAlreadyTaken<br>48: UserLocationOutOfRange<br>49: CategoryNotFound<br>50: CategoryExists<br>51: SubmissionDistanceWrong<br>52: CompanyNotFound<br>53: BranchNotFound<br>54: PhoneAlreadyRegistered<br>55: RoleNotFound<br>56: AlreadyMember<br>57: BuildingNotFound<br>58: ApartmentNotFound<br>59: UnitNotFound<br>60: RoomNotFound<br>61: SignyContactNotFound<br>62: SignyGroupNotFound<br>63: SignyDocumentNotFound<br>64: SignyInputNotFound<br>65: ContactWithSuchEmailExists<br>66: ContactWithSuchPhoneExists<br>67: ContactWithSuchWhatsappExists<br>68: ContactWithSuchTelegramExists<br>69: ContactWithTelegramNickNameExists<br>70: UserNotExists<br>71: SignatoryExists<br>72: SignyEmailTemplateNotFound<br>73: SmsInfoNotFound<br>74: SignatoryNotFound<br>75: PassCodeNotFound<br>76: InvalidPassCode<br>77: InvalidPhoneNumberForIsrael<br>78: SubscriptionPlanNotFound<br>79: CouponNotFound<br>80: SignatoryUserPasswordNotExists<br>81: TemplateNotFound<br>82: UpgradePlanForDocuments<br>83: UpgradePlanForTemplates<br>84: UpgradePlanForSms<br>85: UpgradePlanForStorage<br>86: DocumentAlreadySharedToSignatory"
          },
          "isForceLogout": {
            "type": "boolean",
            "description": "If this parameter is true client app should force user to re-login"
          },
          "isSusbcriptionError": {
            "type": "boolean",
            "description": "If this parameter is true client app should redirect user to subscription screen"
          }
        },
        "required": [
          "message",
          "errorCode",
          "isForceLogout",
          "isSusbcriptionError"
        ]
      },
...
      "UploadedImageInfo": {
        "type": "object",
        "properties": {
          "imageUrl": {
            "type": "string"
          },
          "thumbnailUrl": {
            "type": "string"
          },
          "mimetype": {
            "type": "string"
          }
        }
      },
      "UploadedFileInfo": {
        "type": "object",
        "properties": {
          "mimetype": {
            "type": "string"
          },
          "fileUrl": {
            "type": "string"
          }
        },
        "required": [
          "fileUrl"
        ]
      },
...
      "ApiSuccessResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ]
      },
...
      "CreateEmailTemplateRequest": {
        "type": "object",
        "properties": {
          "emailTemplateId": {
            "type": "number",
            "nullable": true
          },
          "template": {
            "type": "string"
          }
        },
        "required": [
          "template"
        ]
      },
      "EmailTemplateBaseInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "template": {
            "type": "string"
          },
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UploadedFileInfo"
            }
          },
          "companyLogo": {
            "$ref": "#/components/schemas/UploadedImageInfo"
          }
        },
        "required": [
          "id",
          "template"
        ]
      },
      "CreateTemplateResponse": {
        "type": "object",
        "properties": {
          "template": {
            "$ref": "#/components/schemas/EmailTemplateBaseInfo"
          }
        },
        "required": [
          "template"
        ]
      },
      "UploadTemplateFilesToS3Request": {
        "type": "object",
        "properties": {
          "emailTemplateId": {
            "type": "number",
            "nullable": true
          },
          "files": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "binary"
            }
          }
        },
        "required": [
          "files"
        ]
      },
      "UploadTemplateFilesToS3Response": {
        "type": "object",
        "properties": {
          "templateId": {
            "type": "number"
          },
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UploadedFileInfo"
            }
          }
        },
        "required": [
          "templateId",
          "files"
        ]
      },
      "UploadSignyCompanyLogoRequest": {
        "type": "object",
        "properties": {
          "emailTemplateId": {
            "type": "number",
            "nullable": false
          },
          "logo": {
            "type": "string",
            "format": "binary"
          }
        },
        "required": [
          "emailTemplateId",
          "logo"
        ]
      },
      "UploadSignyCompanyLogoResponse": {
        "type": "object",
        "properties": {
          "template": {
            "$ref": "#/components/schemas/EmailTemplateBaseInfo"
          }
        },
        "required": [
          "template"
        ]
      },
      "GetEmailTemplateWithPermissionRequest": {
        "type": "object",
        "properties": {
          "emailTemplateId": {
            "type": "number",
            "nullable": false
          }
        },
        "required": [
          "emailTemplateId"
        ]
      },
      "GetSignyEmailTemplateByIdResponse": {
        "type": "object",
        "properties": {
          "template": {
            "$ref": "#/components/schemas/EmailTemplateBaseInfo"
          }
        },
        "required": [
          "template"
        ]
      },
      "SignySendEmailRequest": {
        "type": "object",
        "properties": {
          "emailTemplateId": {
            "type": "number",
            "nullable": false
          },
          "signatoriesIds": {
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "emailTemplateId",
          "signatoriesIds"
        ]
      },
      "SignySendSmsRequest": {
        "type": "object",
        "properties": {
          "smsTemplateId": {
            "type": "number",
            "nullable": false
          },
          "signatoriesIds": {
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "smsTemplateId",
          "signatoriesIds"
        ]
      },
      "SendSmsResponse": {
        "type": "object",
        "properties": {}
      },
      "GetSmsTemplateWithPermissionRequest": {
        "type": "object",
        "properties": {
          "smsTemplateId": {
            "type": "number",
            "nullable": false
          }
        },
        "required": [
          "smsTemplateId"
        ]
      },
      "SmsBaseInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "message": {
            "type": "string"
          },
          "originator": {
            "type": "string"
          },
          "type": {
            "type": "number",
            "enum": [
              {
                "PhoneVerification": "PHONE_VERIFICATION",
                "ResetPassword": "RESET_PASSWORD",
                "ChangePhone": "CHANGE_PHONE",
                "SignySms": "SIGNY_SMS"
              }
            ]
          }
        },
        "required": [
          "id",
          "message",
          "originator",
          "type"
        ]
      },
      "GetSignySmsTemplateByIdResponse": {
        "type": "object",
        "properties": {
          "template": {
            "$ref": "#/components/schemas/SmsBaseInfo"
          }
        },
        "required": [
          "template"
        ]
      },
      "CreateSmsTemplateRequest": {
        "type": "object",
        "properties": {
          "smsTemplateId": {
            "type": "number",
            "nullable": true
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
      "CreateSmsTemplateResponse": {
        "type": "object",
        "properties": {
          "template": {
            "$ref": "#/components/schemas/SmsBaseInfo"
          }
        },
        "required": [
          "template"
        ]
      },
...
    }
  }
}